requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
pyserial>=3.5
RPi.GPIO>=0.7.0
Adafruit_DHT>=1.4.0
Adafruit_ADS1x15>=1.0.2
SQLAlchemy>=1.4.0
pytest>=6.2.0
python-dotenv>=0.19.0
requests>=2.26.0
scikit-learn>=0.24.0
tensorflow>=2.6.0
opencv-python>=4.5.0
flask>=2.0.0
flask-cors>=3.0.0
qrcode>=7.3.1
Pillow>=8.3.0
fastapi==0.68.1
uvicorn==0.15.0
pydantic==1.8.2
pytest==6.2.5
websockets==10.0
python-multipart==0.0.5 